09210415 已完成api 剩前端上架畫面

+++ FinalS1/src/main/java/tw/Final/FinalS1/repository/SizeRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import tw.Final.FinalS1.model.Size;
 
 @Repository
-public interface SizeRepository extends JpaRepository<Size, Integer> {
+public interface SizeRepository extends JpaRepository<Size, Long> {
 }
